user_mgmt_service/requirements.txt
fastapi==0.109.0
uvicorn==0.29.0
sqlalchemy==1.4.40
psycopg2-binary==2.9.7
pydantic==1.10.12
httpx==0.27.0
kafka-python==2.0.2